Backend/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
flask-mongoengine==1.0.0
python-dotenv==1.0.0
mongoengine==0.24.0
dnspython==2.3.0
pymongo==4.3.1
PyJWT==2.8.0
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.1
Flask-Limiter==3.3.1
limits==3.3.1
redis==4.5.4
pyOpenSSL==23.2.0
gunicorn==21.2.0
flask-cors==4.0.0
flask-marshmallow==0.15.0
marshmallow-mongoengine==0.31.1
flasgger==0.9.7
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
